make EditorViewController makeWindow, added on closed document
