+++ VoiceOver Designer/VODocumentController.swift
@@ -10,26 +10,29 @@ import Foundation
 import AppKit
 import Document
 import Editor
+import Projects
 
 class VODocumentController: NSDocumentController {
     override func documentClass(forType typeName: String) -> AnyClass? {
         return VODesignDocument.self
     }
+    
+    override func removeDocument(_ document: NSDocument) {
+        super.removeDocument(document)
+        
+        if VODocumentController.shared.documents.isEmpty {
+            let controller = ProjectsViewController.fromStoryboard()
+            let wc = WindowContoller.fromStoryboard()
+            wc.window?.setContentSize(.init(width: 800, height: 400))
+            wc.window?.toolbar = controller.toolbar
+            wc.window?.setFrameAutosaveName("windowFrame")
+            wc.showWindow(self)
+        }
+    }
 }
 
 extension VODesignDocument {
     public override func makeWindowControllers() {
-        let controller = EditorViewController.fromStoryboard()
-        controller.presenter.document = self
-        let window = NSWindow(contentViewController: controller)
-        window.setContentSize(NSSize(width: 800, height: 600))
-        
-        let wc = NSWindowController(window: window)
-        wc.contentViewController = controller
-        wc.window?.toolbar = controller.toolbar
-        addWindowController(wc)
-        
-        window.setFrameAutosaveName("windowFrame")
-        window.makeKeyAndOrderFront(self)
+        EditorViewController.makeWindow(for: self)
     }
 }
